Remove custom font
As explained in #179, the custom font is unnecessary for our situation and removing it could help improving page load significantly.

Note that removing our option means that the default font choice of Bootstrap will be used. Since we are on version 4, it will be a [native font stack](https://getbootstrap.com/docs/4.5/content/reboot/#native-font-stack) which I think is really good for our case because there would be no loading or fetching at all.